+++ count_kmer/src/count_lib.c
@@ -157,7 +157,7 @@ void my_error(void *ptr, const char *erreur)
 }
 
 
-void count(const char *seq,const size_t seq_taille,int *tab_bool,int *taille_kmer,int **resultat,int taille_sous_sequence)
+void count(const char *seq,const size_t seq_taille,int *tab_bool,int *taille_kmer,unsigned long long **resultat,int taille_sous_sequence)
 {
 
 
@@ -231,7 +231,7 @@ void count(const char *seq,const size_t seq_taille,int *tab_bool,int *taille_kme
 
 
 
-void add_one(const char *seq,int i,const size_t seq_taille,int pattern_taille,int *resultat,int *tab_bool,int kmer_taille)
+void add_one(const char *seq,int i,const size_t seq_taille,int pattern_taille,unsigned long long *resultat,int *tab_bool,int kmer_taille)
 {
 	int alpha =  i + seq_taille - (pattern_taille + 1);
 	int beta = i + seq_taille - 1 ;
@@ -286,7 +286,7 @@ void add_one(const char *seq,int i,const size_t seq_taille,int pattern_taille,in
  * @param kmer_taille: la taille du kmer
  * @param tab_comptage: la table de comptage
  * */
-void count_seq(const char *seq, const size_t seq_taille, int *tab_bool, int *resultat,int pattern_taille) 
+void count_seq(const char *seq, const size_t seq_taille, int *tab_bool, unsigned long long *resultat,int pattern_taille) 
 {
 	long long position;
 	long long i;
